js/no-js CSS & SVG title update
if JavaScript is blocked, or unavailable, this commi will ensure that content is still viewable by making fade-in transitions reliant on the HTML element having a class of "js".

If JavaScript is disabled, the no-js class stayes on the HTML element, and the animated SVG "Open Source" is hidden from the DOM and the normally sr-only heading is brought into view as a legible replacement.

This commit also takes care of a significant AT bug with the SVG, where "open source" was previously announced 6x in a row, likely due to the animation effect.  By adding a sr-only h1 to the area, and forcing AT to recognize the svg as a role="img" with a defined aria-label, it announces a more expected result.